refactor: Move StatusTypes to BridgeController
